update build scripts

+++ build.sh
@@ -1,17 +1,25 @@
 #!/bin/bash
-if [[ "$1" == "simulator" ]];then
-    cargo build --release --example $2 --target x86_64-apple-darwin
-    out=target/x86_64-apple-darwin/release/examples
-    mkdir -p $out/source
-    touch $out/source/pdex.bin
-    cp $out/lib$2.dylib $out/source/pdex.dylib
-    mkdir -p build
-    pdc $out/source build/$2.pdx
-elif [[ "$1" == "device" ]];then
-    cargo build --release --example $2 --target thumbv7em-none-eabihf
-    out=target/thumbv7em-none-eabihf/release/examples
-    mkdir -p $out/source
-    arm-none-eabi-objcopy -O binary $out/$2 $out/source/pdex.bin
-    mkdir -p build
-    pdc $out/source build/$2-device.pdx
+if [[ "$1" == "clean" ]] || if [[ "$1" == "simulator" ]] || if [[ "$1" == "device" ]];then
+    if [[ "$1" == "clean" ]];then
+        cargo clean
+        rm -r build
+    else
+        if [[ "$1" == "simulator" ]];then
+            target=x86_64-apple-darwin
+        elif [[ "$1" == "device" ]];then
+            target=thumbv7em-none-eabihf
+        fi
+        cargo build --release --example $2 --target $target
+        out=target/$target/release/examples
+        mkdir -p $out/source
+        mkdir -p build
+        if [[ "$1" == "simulator" ]];then
+            touch $out/source/pdex.bin
+            cp $out/lib$2.dylib $out/source/pdex.dylib
+            pdc $out/source build/$2.pdx
+        elif [[ "$1" == "device" ]];then
+            arm-none-eabi-objcopy -O binary $out/$2 $out/source/pdex.bin
+            pdc $out/source build/$2-device.pdx
+        fi
+    fi
 fi
